refactor: UserBadge 디렉토리 Comment 디렉토리 하위로 이동
