+++ components/comment/userBadge/IndicatorIcon.tsx
