+++ components/comment/userBadge/UserBadgeIndicator.tsx
@@ -4,10 +4,10 @@ import IndicatorIcon from './IndicatorIcon';
 import EditModal from './modal/EditModal';
 import DeleteModal from './modal/DeleteModal';
 import IndicatorPopover from './IndicatorPopover';
-import usePopover from '../../hooks/usePopover';
-import useModal from '../../hooks/useModal';
-import useComment from '../../hooks/useComment';
-import { commentRefetchKeyAtom } from '../../store/comment';
+import usePopover from '../../../hooks/usePopover';
+import useModal from '../../../hooks/useModal';
+import useComment from '../../../hooks/useComment';
+import { commentRefetchKeyAtom } from '../../../store/comment';
 
 interface UserBadgeIndicatorProps {
   depth: 1 | 2;
